backend/pdfs/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pymongo==4.6.0
PyPDF2==3.0.1
python-multipart==0.0.6
pydantic==2.5.0
python-dotenv==1.0.0
pymupdf==1.26.0
Pillow==10.1.0
pdf2image==1.16.3
httpx==0.28.1